requirements.txt
openai==0.28.1
streamlit==1.36.0
python-dotenv==1.0.1
